refactor(api): mission email shared dto
