+++ packages/dto/src/resources/index.ts
@@ -1,3 +1,4 @@
 export * from "./mission-browse";
+export * from "./mission-email";
 export * from "./mission-match";
 export * from "./user-scoring";
