+++ plateform/app/services/email.ts
@@ -1,18 +1,7 @@
+import type { SendMissionEmailRequest, SendMissionEmailResponse } from "@engagement/dto";
+
 import { client } from "~/services/client";
 
-export type SendMissionEmailPayload = {
-  email: string;
-  publisherId: string;
-  userScoringId?: string;
-  distinctId?: string;
-  missionIds?: string[];
-};
-
-export type SendMissionEmailResponse = {
-  email_sent: boolean;
-  email_skip_reason?: string;
-};
-
-export async function sendMissionEmail(payload: SendMissionEmailPayload): Promise<SendMissionEmailResponse> {
+export async function sendMissionEmail(payload: SendMissionEmailRequest): Promise<SendMissionEmailResponse> {
   return client.post<SendMissionEmailResponse>("/api/email/mission", payload);
 }
